+++ src/Aws/DataPipeline/Resources/datapipeline-2012-10-29.php
@@ -120,6 +120,77 @@ return array (
                 ),
             ),
         ),
+        'AddTags' => array(
+            'httpMethod' => 'POST',
+            'uri' => '/',
+            'class' => 'Aws\\Common\\Command\\JsonCommand',
+            'responseClass' => 'EmptyOutput',
+            'responseType' => 'model',
+            'parameters' => array(
+                'Content-Type' => array(
+                    'static' => true,
+                    'location' => 'header',
+                    'default' => 'application/x-amz-json-1.1',
+                ),
+                'command.expects' => array(
+                    'static' => true,
+                    'default' => 'application/json',
+                ),
+                'X-Amz-Target' => array(
+                    'static' => true,
+                    'location' => 'header',
+                    'default' => 'DataPipeline.AddTags',
+                ),
+                'pipelineId' => array(
+                    'required' => true,
+                    'type' => 'string',
+                    'location' => 'json',
+                    'minLength' => 1,
+                    'maxLength' => 1024,
+                ),
+                'tags' => array(
+                    'required' => true,
+                    'type' => 'array',
+                    'location' => 'json',
+                    'maxItems' => 10,
+                    'items' => array(
+                        'name' => 'Tag',
+                        'type' => 'object',
+                        'properties' => array(
+                            'key' => array(
+                                'required' => true,
+                                'type' => 'string',
+                                'minLength' => 1,
+                                'maxLength' => 128,
+                            ),
+                            'value' => array(
+                                'required' => true,
+                                'type' => 'string',
+                                'maxLength' => 256,
+                            ),
+                        ),
+                    ),
+                ),
+            ),
+            'errorResponses' => array(
+                array(
+                    'reason' => 'An internal service error occurred.',
+                    'class' => 'InternalServiceErrorException',
+                ),
+                array(
+                    'reason' => 'The request was not valid. Verify that your request was properly formatted, that the signature was generated with the correct credentials, and that you haven\'t exceeded any of the service limits for your account.',
+                    'class' => 'InvalidRequestException',
+                ),
+                array(
+                    'reason' => 'The specified pipeline was not found. Verify that you used the correct user and account identifiers.',
+                    'class' => 'PipelineNotFoundException',
+                ),
+                array(
+                    'reason' => 'The specified pipeline has been deleted.',
+                    'class' => 'PipelineDeletedException',
+                ),
+            ),
+        ),
         'CreatePipeline' => array(
             'httpMethod' => 'POST',
             'uri' => '/',
@@ -159,6 +230,28 @@ return array (
                     'type' => 'string',
                     'location' => 'json',
                     'maxLength' => 1024,
+                ),
+                'tags' => array(
+                    'type' => 'array',
+                    'location' => 'json',
+                    'maxItems' => 10,
+                    'items' => array(
+                        'name' => 'Tag',
+                        'type' => 'object',
+                        'properties' => array(
+                            'key' => array(
+                                'required' => true,
+                                'type' => 'string',
+                                'minLength' => 1,
+                                'maxLength' => 128,
+                            ),
+                            'value' => array(
+                                'required' => true,
+                                'type' => 'string',
+                                'maxLength' => 256,
+                            ),
+                        ),
+                    ),
                 ),
             ),
             'errorResponses' => array(
@@ -799,6 +892,64 @@ return array (
                 ),
             ),
         ),
+        'RemoveTags' => array(
+            'httpMethod' => 'POST',
+            'uri' => '/',
+            'class' => 'Aws\\Common\\Command\\JsonCommand',
+            'responseClass' => 'EmptyOutput',
+            'responseType' => 'model',
+            'parameters' => array(
+                'Content-Type' => array(
+                    'static' => true,
+                    'location' => 'header',
+                    'default' => 'application/x-amz-json-1.1',
+                ),
+                'command.expects' => array(
+                    'static' => true,
+                    'default' => 'application/json',
+                ),
+                'X-Amz-Target' => array(
+                    'static' => true,
+                    'location' => 'header',
+                    'default' => 'DataPipeline.RemoveTags',
+                ),
+                'pipelineId' => array(
+                    'required' => true,
+                    'type' => 'string',
+                    'location' => 'json',
+                    'minLength' => 1,
+                    'maxLength' => 1024,
+                ),
+                'tagKeys' => array(
+                    'required' => true,
+                    'type' => 'array',
+                    'location' => 'json',
+                    'items' => array(
+                        'name' => 'string',
+                        'type' => 'string',
+                        'maxLength' => 1024,
+                    ),
+                ),
+            ),
+            'errorResponses' => array(
+                array(
+                    'reason' => 'An internal service error occurred.',
+                    'class' => 'InternalServiceErrorException',
+                ),
+                array(
+                    'reason' => 'The request was not valid. Verify that your request was properly formatted, that the signature was generated with the correct credentials, and that you haven\'t exceeded any of the service limits for your account.',
+                    'class' => 'InvalidRequestException',
+                ),
+                array(
+                    'reason' => 'The specified pipeline was not found. Verify that you used the correct user and account identifiers.',
+                    'class' => 'PipelineNotFoundException',
+                ),
+                array(
+                    'reason' => 'The specified pipeline has been deleted.',
+                    'class' => 'PipelineDeletedException',
+                ),
+            ),
+        ),
         'ReportTaskProgress' => array(
             'httpMethod' => 'POST',
             'uri' => '/',
@@ -1318,6 +1469,21 @@ return array (
                             ),
                             'description' => array(
                                 'type' => 'string',
+                            ),
+                            'tags' => array(
+                                'type' => 'array',
+                                'items' => array(
+                                    'name' => 'Tag',
+                                    'type' => 'object',
+                                    'properties' => array(
+                                        'key' => array(
+                                            'type' => 'string',
+                                        ),
+                                        'value' => array(
+                                            'type' => 'string',
+                                        ),
+                                    ),
+                                ),
                             ),
                         ),
                     ),